requirements.txt
hydra-core~=1.1.0
pytorch-lightning~=1.3.8
pytorch-lightning-bolts
transformers~=4.8.2
torch~=1.9.0
torchmetrics~=0.5.1
omegaconf~=2.1.1